+++ models/src/main/java/io/sphere/sdk/products/commands/updateactions/RemovePrice.java
@@ -17,7 +17,7 @@ public class RemovePrice extends StageableProductUpdateAction {
         this.price = price;
     }
 
-    public long getVariantId() {
+    public int getVariantId() {
         return variantId;
     }
